deploy/requirements.txt
doit
PyYAML
